% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPCA.R
\name{cPCA}
\alias{cPCA}
\title{Contrastive PCA}
\usage{
cPCA(Xt, Xa, eta, sparse_ctst = NULL, sparse_trt = NULL, sparse_ctrl = NULL)
}
\arguments{
\item{Xt}{data frame for target group, samples in rows, variables in columns.}

\item{Xa}{data frame for ancillary group, samples in rows, variables in columns.}

\item{eta}{tuning parameter controlling how much ancillary variation should be contrasted off from the target variation. Typically the \code{eta_opt} returned by \code{\link{eta_tuning_general}}.}

\item{sparse_ctst}{parameter controlling the sparsity of scPCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_trt}{parameter controlling the sparsity of target sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_ctrl}{parameter controlling the sparsity of ancillary sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}
}
\value{
A list containing contrastive scores \code{cPC} and loadings \code{cU}, target scores \code{PCt} and loadings \code{Ut}, and ancillary scores \code{PCa} and loadings \code{Ua}.
}
\description{
This function performs contrastive PCA (cPCA) or sparse contrastive PCA (scPCA), in comparison, also perform target/ancillary group only PCA/sPCA
}
\examples{
library(ConSpec)
data(intro)
cpca_res <- cPCA(Xt = intro$cPCA$Xt, Xa = intro$cPCA$Xa, eta = 1.2, sparse_ctst = NULL, sparse_trt = NULL, sparse_ctrl = NULL)
}
