% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_tuning_general.R
\name{eta_tuning_general}
\alias{eta_tuning_general}
\title{Select \eqn{\eta} for contrastive spectral clustering by cPCA or scPCA}
\usage{
eta_tuning_general(
  Xt,
  Xa,
  eta_range = seq(0.5, 10, 0.5),
  plot = TRUE,
  sparse_trt = NULL,
  sparse_ctst = NULL,
  num_comps = 20,
  km_cluster = 2,
  ckm_cluster = 2,
  clust_comps = "auto",
  nstart = 25
)
}
\arguments{
\item{Xt}{data frame for target group, samples in rows, variables in columns. If PCA is applied to correlation matrix, Xt should be scaled and centered.}

\item{Xa}{data frame for ancillary group, samples in rows, variables in columns. If cPCA is applied to correlation matrix, Xa should also be scaled and centered.}

\item{eta_range}{a vector containing candidate eta range to be tuning from.}

\item{plot}{a logical quantity indicating whether plot the ARI(nuisance) and silhouette score.}

\item{sparse_trt}{parameter controlling the sparsity of target sparse PCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{sparse_ctst}{parameter controlling the sparsity of scPCA, if NULL, use cPCA, else use scPCA. For use of scPCA, if vector, refer sumabss in PMD.cv from package PMA, if single number, refer sumabs in PMD in package PMA.}

\item{num_comps}{if scPCA, specify the number of contrastive components that will be estimated by PMD, default is 20.}

\item{km_cluster}{number showing number of nuisance clusters defined from kmeans clustering on target PCA}

\item{ckm_cluster}{number of interesting clusters defined from kmeans clustering on cPCA/scPCA}

\item{clust_comps}{number of leading component scores used for the kmeans clustering and the silhouette score, for both the target-only and the contrastive analysis. The default \code{"auto"} uses \code{ckm_cluster - 1}, the smallest number of score dimensions that can separate \code{ckm_cluster} clusters. A numeric value overrides the rule. Using more components than \code{ckm_cluster - 1} allows residual nuisance variation to re-enter the criterion and inflates the selected \eqn{\eta}.}

\item{nstart}{number of random starts passed to \code{\link[stats]{kmeans}}.}
}
\value{
A list containing selected value of \eqn{\eta}  \code{eta_opt}, the vector of ARI(nuisance) \code{nuisance_ARI} and the vector of Silhouette \code{silhouette}, and the ggplot object \code{p}.
}
\description{
Select \eqn{\eta} for contrastive spectral clustering by cPCA or scPCA
}
\examples{
library(ConSpec)
# A simple and balanced data example
data(intro)
eta_range <- seq(0.1, 5, 0.05)
eta_opt <- eta_tuning_general(Xt = intro$cPCA$Xt, Xa = intro$cPCA$Xa, eta_range = eta_range, plot = TRUE, sparse_trt = NULL, sparse_ctst = NULL)$eta_opt
# An unbalanced data example
data(sim_data_CSC)
eta_tuning_general(Xt = sim_data_CSC$Xt, Xa = sim_data_CSC$Xa, eta_range = seq(0, 100, 0.2))
}
