% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sCSC.R
\name{sim_sCSC}
\alias{sim_sCSC}
\title{Simulation function for contrastive spectral clustering (CSC)}
\usage{
sim_sCSC(
  n11,
  n12,
  n21,
  n22,
  m1tilde,
  m2tilde,
  mu1,
  mu2,
  theta1,
  theta2,
  p1,
  p2,
  p3,
  sigma_t = 1,
  sigma_a = 1,
  Sigma_t = NULL,
  Sigma_a = NULL,
  eta,
  repetition = 10
)
}
\arguments{
\item{n11}{number of target samples in interesting cluster 1 and nuisance cluster 1.}

\item{n12}{number of target samples in interesting cluster 1 and nuisance cluster 2.}

\item{n21}{number of target samples in interesting cluster 2 and nuisance cluster 1.}

\item{n22}{number of target samples in interesting cluster 2 and nuisance cluster 2.}

\item{m1tilde}{number of ancillary samples in nuisance cluster 1.}

\item{m2tilde}{number of ancillary samples in nuisance cluster 2.}

\item{mu1}{non-zero values in mu_t1.}

\item{mu2}{non-zero values in mu_t2.}

\item{theta1}{non-zero values in theta_1.}

\item{theta2}{non-zero values in theta_2}

\item{p1}{number of non-zero elements in mu_t1 and mu_t2.}

\item{p2}{number of non-zero elements in theta_1 and theta_2.}

\item{p3}{number of zero elements.}

\item{sigma_t}{standard deviation of the uncorrelated errors in target group.}

\item{sigma_a}{standard deviation of the uncorrelated errors in ancillary group.}

\item{Sigma_t}{the error covariance matrix in target group, if is NULL, use standard deviation specified and assume uncorrelation.}

\item{Sigma_a}{the error covariance matrix in ancillary group, if is NULL, use standard deviation specified and assume uncorrelation.}

\item{eta}{if eta is a vector, eta_tuning algorithm will be called to select optimal eta; if a number, eta will be used.}

\item{repetition}{number of replications of the simulation.}
}
\value{
A data frame containing the selected optimal eta, the four metrics values.
}
\description{
This function performs simulation to prove the interesting of hidden clusters by using cPCA and eta tuning algorithm.
Same as in the example of Section 2, we consider two target clusters and two nuisance clusters
}
\examples{
# See vignettes/ContrastiveSpectralClustering.Rmd for example usage.
}
